changed assignment code

+++ src/main/kotlin/dev/leonzimmermann/demo/extendablespringdemo/services/AssignmentService.kt
@@ -4,5 +4,5 @@ import dev.leonzimmermann.demo.extendablespringdemo.models.Assignment
 
 interface AssignmentService {
   fun generateNewAssignment(): Assignment
-  fun solveAssignmentAndReturnListOfDiscrepancies(objectId: Int, answer: String): List<String>
+  fun solveAssignmentAndReturnListOfDiscrepancies(objectId: Long, answer: String): List<String>
 }
